add column

+++ app/Http/Controllers/TasksController.php
@@ -15,16 +15,24 @@ class TasksController extends Controller
      */
     public function index()
     {
-        if(\Auth::check()){
-            $tasks = Task::all();
-            return view('tasks.index', [
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
                 'tasks' => $tasks,
-            ]);  
-        }else{
-            return  view('welcome');  
+            ];
         }
-        
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
+        
+    
 
     /**
      * Show the form for creating a new resource.
@@ -53,11 +61,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
-        
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            
+        ]);
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 
@@ -105,11 +116,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-         $task = Task::findOrFail($id);
-         $task->status = $request->status; 
-         $task->content = $request->content;
-        $task->save();
-        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            
+        ]);
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 
@@ -121,8 +135,17 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::findOrFail($id);
-         $task->delete();
-         return redirect('/');
+         // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        return redirect('/');
+    
+         
     }
 }
